server/requirements.txt
flask 
transformers 
torch 
pdf2image 
pillow 
pytesseract 
requests 
flask 
python-dotenv 
flask-limiter 
PyPDF2
werkzeug
uuid
re